fix(types): correct import paths after modularizing interfaces
